+++ src/main/storage/sync/errors.ts
@@ -1,0 +1,6 @@
+export class RemoteSnapshotPendingError extends Error {
+  constructor(message = "Remote snapshot is still downloading from iCloud") {
+    super(message)
+    this.name = "RemoteSnapshotPendingError"
+  }
+}